webkit-efl/LayoutTests/platform/efl-slp/media/video-timeupdate-during-playback-expected.txt

CONSOLE MESSAGE: line 27: TypeError: 'undefined' is not a function (evaluating 'element.canPlayType(codecs[i][0])')
FAIL: Timed out waiting for notifyDone to be called
Test 'timeupdate' events are posted while playing but not while paused.
